Урок 5. ДЗ. Предварительная подготовка к выполнению ДЗ: float -> double

+++ app/src/main/java/ru/lyubimov/weather/weatherapp/model/DatabaseWeatherModel.kt
@@ -2,8 +2,8 @@ package ru.lyubimov.weather.weatherapp.model
 
 data class DatabaseWeatherModel(
         var city: String? = null,
-        var lat: Float? = 0f,
-        var lon: Float? = 0f,
+        var lat: Double? = 0.0,
+        var lon: Double? = 0.0,
         //т. к. первоначально приложение не проектировалось для использования с базой данных и первоначальная
         //архитектура - жесткая, без следования принципам SOLID, будем костылить: List<Weather> сохраним как
         //Json строку и при сохранении и чтении из базы будем конвертировать туда-обратно.
