+++ app/src/main/java/ru/lyubimov/weather/weatherapp/model/RequestWeatherModel.kt
@@ -17,9 +17,9 @@ class Weather {
 }
 
 class Temperature {
-    @Expose @SerializedName("temp") var temp: Float = 0f
-    @Expose @SerializedName("temp_min") var tempMin: Float = 0f
-    @Expose @SerializedName("temp_max") var tempMax: Float = 0f
+    @Expose @SerializedName("temp") var temp: Double = 0.0
+    @Expose @SerializedName("temp_min") var tempMin: Double = 0.0
+    @Expose @SerializedName("temp_max") var tempMax: Double = 0.0
 }
 
 class Condition {
@@ -42,6 +42,6 @@ class City {
 }
 
 class Coordinate {
-    @Expose @SerializedName("lat") var latitude: Float? = 0f
-    @Expose @SerializedName("lon") var longitude: Float? = 0f
+    @Expose @SerializedName("lat") var latitude: Double? = 0.0
+    @Expose @SerializedName("lon") var longitude: Double? = 0.0
 }
